// scrivafowcl/src/app/shared/components/form-inputs/scriva-typeahead/utilities/scriva-typeahead.interfaces.ts
/*
 * ========================LICENSE_START=================================
 * 
 * Copyright (C) 2025 Regione Piemonte
 * 
 * SPDX-FileCopyrightText: (C) Copyright 2025 Regione Piemonte
 * SPDX-License-Identifier: EUPL-1.2
 * =========================LICENSE_END==================================
 */
import { ScrivaFormInputTypeahead } from '../../form-input/utilities/form-input.classes';

export interface ITypeAheadDataValidazioneTmpl {
  dataConfig: ScrivaFormInputTypeahead;
  data: any;
  typeaheadMap: (v: any) => string;
}